test_framework/requirements.txt
Appium-Python-Client
pytest
selenium
allure-pytest